# Optimize dumping objects count